Refactor: Move Template to components folder
