+++ frontend/app/(defaultLayout)/Troubleshooting/[device]/components/pageTemplate.tsx
